Install git binary package from git-core PPA
Ubuntu focal (20.04) has git v2.25, but the minimum needed for GitHub
Actions runners is v2.29.  Because of that, git has been built from
source for this Docker image.

However, to save time, we can use the git-core PPA to get the latest
stable version of git.  In the future, when updating to Ubuntu hirsute
(21.04) or later, we can remove the git-core PPA in favor of the regular
Ubuntu git package.

Closes #125